+++ Sources/KeyboardKit/_Migrations/MigrationTypes.swift
@@ -17,6 +17,9 @@ import SwiftUI
 @available(*, deprecated, renamed: "Callouts.ActionCallout")
 public typealias ActionCallout = Callouts.ActionCallout
 
+@available(*, deprecated, renamed: "InputSet")
+public typealias AlphabeticInputSet = InputSet
+
 @available(*, deprecated, renamed: "AudioFeedback.Configuration")
 public typealias AudioFeedbackConfiguration = AudioFeedback.Configuration
 
@@ -47,6 +50,15 @@ public typealias HapticFeedbackEngine = HapticFeedback.Engine
 @available(*, deprecated, renamed: "Callouts.InputCallout")
 public typealias InputCallout = Callouts.InputCallout
 
+@available(*, deprecated, renamed: "InputSet.Item")
+public typealias InputSetItem = InputSet.Item
+
+@available(*, deprecated, renamed: "InputSet.Row")
+public typealias InputSetRow = InputSet.Row
+
+@available(*, deprecated, renamed: "InputSet.Row")
+public typealias InputSetRows = InputSet.Rows
+
 @available(*, deprecated, renamed: "Gestures.KeyboardGesture")
 public typealias KeyboardGesture = Gestures.KeyboardGesture
 
@@ -59,6 +71,9 @@ public typealias KeyboardEnabledLabel = KeyboardStateLabel
 @available(*, deprecated, renamed: "KeyboardStateInspector")
 public typealias KeyboardEnabledStateInspector = KeyboardStateInspector
 
+@available(*, deprecated, renamed: "InputSet")
+public typealias NumericInputSet = InputSet
+
 @available(*, deprecated, renamed: "Gestures.RepeatTimer")
 public typealias RepeatGestureTimer = Gestures.RepeatTimer
 
@@ -70,6 +85,9 @@ public typealias SpaceDragSensitivity = Gestures.SpaceDragSensitivity
 
 @available(*, deprecated, renamed: "Gestures.SpaceLongPressBehavior")
 public typealias SpaceLongPressBehavior = Gestures.SpaceLongPressBehavior
+
+@available(*, deprecated, renamed: "InputSet")
+public typealias SymbolicInputSet = InputSet
 
 @available(*, deprecated, renamed: "KeyboardButton.Button")
 public typealias SystemKeyboardButton = KeyboardButton.Button