+++ Tests/KeyboardKitTests/Input/InputSet+ItemTests.swift
@@ -1,5 +1,5 @@
 //
-//  InputSetItemTests.swift
+//  InputSet+ItemTests.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2020-07-03.
@@ -9,10 +9,10 @@
 import KeyboardKit
 import XCTest
 
-class InputSetItemTests: XCTestCase {
+class InputSet_ItemTests: XCTestCase {
     
     func testCharacterForCasingReturnsCorrectChar() {
-        let input = InputSetItem(neutral: "n", uppercased: "u", lowercased: "l")
+        let input = InputSet.Item(neutral: "n", uppercased: "u", lowercased: "l")
         XCTAssertEqual(input.character(for: .auto), "l")
         XCTAssertEqual(input.character(for: .lowercased), "l")
         XCTAssertEqual(input.character(for: .uppercased), "u")